// apps/dashboard/src/interfaces/deployment.ts
export type RuntimeMode = 'web' | 'worker';
export type BuildEngine =
  | 'nixpacks'
  | 'dockerfile'
  | 'buildpacks'
  | 'static'
  | 'railpack'
  | 'serverless';
export type ServiceStatus = 'created' | 'building' | 'running' | 'stopped' | 'error';
// ...
export interface AppService {
  id: string;
  projectId: string;
  environmentId: string;
  name: string;
  repositoryUrl: string;
  imageRef?: string;
  branch: string;
  rootDirectory: string;
  runtimeMode: RuntimeMode;
  installCommand: string;
  buildCommand: string;
  startCommand: string;
  dockerfilePath: string;
  buildEngine: BuildEngine;
  internalPort: number;
  domain: string;
  staticOutput: string;
  healthCheckPath: string;
  containerId: string;
  status: ServiceStatus;
  createdAt: string;
  updatedAt: string;
  cpuLimit?: number;
  memoryLimit?: number;
  enablePRPreviews: boolean;
  maintenanceMode: boolean;
  icon?: string;
  deployToken?: string;
}